refactor/feat: new design based on verifiers + appcompose verification
The new design allow supporting new verifiers based on different TEE
techs. There are more ways to generalize, including unifying the quote
format, but that's more than we need right now. The main purpose of the
new design is to configure a verifier at the start with different
attributes like the docker-compose that should be running in the TEE.